SHA-133: Refactor file names to kebab case (#276)
