+++ src/js/comment-box/content-script.ts
@@ -1,4 +1,4 @@
-import {logError} from '../utils/logError'
+import {logError} from '../utils/log-error'
 import sleep from '../utils/sleep'
