+++ src/js/content-scripts.ts
@@ -3,12 +3,12 @@ import * as Sentry from '@sentry/browser'
 import {analyzeStoryDescription} from './analyze/analyze-story-description'
 import {CycleTime} from './cycle-time/cycle-time'
 import {DevelopmentTime} from './development-time/development-time'
-import {KeyboardShortcuts} from './keyboard/keyboardShortcuts'
-import {NotesButton} from './notes/notesButton'
-import {Todoist} from './todoist/Todoist'
-import {getSyncedSetting} from './utils/getSyncedSetting'
-import {logError} from './utils/logError'
-import storyPageIsReady from './utils/storyPageIsReady'
+import {KeyboardShortcuts} from './keyboard/keyboard-shortcuts'
+import {NotesButton} from './notes/notes-button'
+import {Todoist} from './todoist/todoist'
+import {getSyncedSetting} from './utils/get-synced-setting'
+import {logError} from './utils/log-error'
+import storyPageIsReady from './utils/story-page-is-ready'
 
 
 const manifestData = chrome.runtime.getManifest()
