+++ src/js/popup/popup.ts
@@ -1,7 +1,7 @@
 import * as Sentry from '@sentry/browser'
 
 import {sendEvent} from '../analytics/event'
-import {getSyncedSetting} from '../utils/getSyncedSetting'
+import {getSyncedSetting} from '../utils/get-synced-setting'
 import sleep from '../utils/sleep'
 
 import {NotesPopup} from './notes-popup'
