+++ src/js/service-worker/service-worker.ts
@@ -3,8 +3,8 @@ import * as Sentry from '@sentry/browser'
 import callOpenAI from '../ai/call-openai'
 import getOpenAiToken from '../ai/get-openai-token'
 import {sendEvent} from '../analytics/event'
-import {getActiveTab} from '../utils/getActiveTab'
-import {getSyncedSetting} from '../utils/getSyncedSetting'
+import {getActiveTab} from '../utils/get-active-tab'
+import {getSyncedSetting} from '../utils/get-synced-setting'
 import {Story} from '../utils/story'
 
 import {onInstallAndUpdate} from './on-install-and-update'
