+++ src/js/utils/get-synced-setting.ts
