+++ tests/service-worker/omnibox/redirect-from-omnibox.test.js
