+++ tests/utils/get-active-tab.test.js
@@ -1,4 +1,4 @@
-import {getActiveTab} from '../../src/js/utils/getActiveTab'
+import {getActiveTab} from '../../src/js/utils/get-active-tab'
 
 
 describe('getActiveTab', () => {
